requirements.txt
openai>=1.58.0
pyaudio>=0.2.14
python-dotenv>=1.0.0
